requirements.txt
opencv-python
numpy
matplotlib
streamlit
scikit-learn
pillow